devide input_url_widget and crawling_image_widget
