+++ lib/app_route.dart
@@ -1,5 +1,6 @@
 class AppRoute {
   static const String home = "/";
   static const String pinDetail = "/pin_detail";
+  static const String inputUrl = "/input_url";
   static const String crawlingImage = "/crawling_image";
 }
