+++ lib/main.dart
@@ -6,6 +6,7 @@ import 'package:pintersest_clone/api/pins_api.dart';
 import 'package:pintersest_clone/app_route.dart';
 import 'package:pintersest_clone/data/pins_repository.dart';
 import 'package:pintersest_clone/view/main/crawling_image/crawling_image_widget.dart';
+import 'package:pintersest_clone/view/main/input_url_widget/input_url_widget.dart';
 import 'package:pintersest_clone/view/main/main_navigation_page.dart';
 import 'package:pintersest_clone/view/main/pin_detail_widget/pin_detail_widget.dart';
 
@@ -41,6 +42,7 @@ class MyApp extends StatelessWidget {
           routes: {
             AppRoute.home: (context) => MainNavigationPage(),
             AppRoute.pinDetail: (context) => PinDetailWidget(),
+            AppRoute.inputUrl: (context) => InputUrlWidget(),
             AppRoute.crawlingImage: (context) => CrawlingImageWidget(),
           }),
     );
